+++ eras/dijkstra/impl/testlib/Test/Cardano/Ledger/Dijkstra/ImpTest.hs
@@ -22,6 +22,8 @@ module Test.Cardano.Ledger.Dijkstra.ImpTest (
   fixupSubTransactions,
   balanceSubTransactions,
   switchTxToLegacyMode,
+  txWithSubTxs,
+  submitFailingSubTx,
 ) where
 
 import Cardano.Ledger.Allegra.Scripts (
@@ -147,6 +149,24 @@ instance InjectRuleFailure "LEDGER" DijkstraSubUtxoPredFailure DijkstraEra where
       . SubLedgerFailure
       . SubUtxowFailure
       . SubUtxoFailure
+
+-- | An otherwise empty top level transaction carrying the given
+-- sub-transactions.
+txWithSubTxs :: DijkstraEraImp era => [Tx SubTx era] -> Tx TopTx era
+txWithSubTxs subTxs =
+  mkBasicTx mkBasicTxBody & bodyTxL . subTransactionsTxBodyL .~ OMap.fromFoldable subTxs
+
+-- | Submit a sub-transaction, nested in an otherwise empty top level
+-- transaction, that is expected to be rejected with exactly the given
+-- predicate failures.
+submitFailingSubTx ::
+  ( HasCallStack
+  , DijkstraEraImp era
+  ) =>
+  Tx SubTx era ->
+  NonEmpty (PredicateFailure (EraRule "LEDGER" era)) ->
+  ImpTestM era ()
+submitFailingSubTx subTx = submitFailingTx $ txWithSubTxs [subTx]
 
 impDijkstraSatisfyNativeScript ::
   ( DijkstraEraImp era